# readme/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/scenes/{sceneId}/events": {
      "get": {
        "operationId": "ScenesSSEController_streamSceneEvents",
        "summary": "Connect to scene events via Server-Sent Events (SSE)",
        "description": "Fallback for realtime scene updates when WebSockets are not available",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sceneId",
            "required": true,
            "in": "path",
            "description": "Scene ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Scenes SSE"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumea API",
    "description": "# Lumea Interior Layout Generator API\n\nAdvanced AI-powered interior layout generator with explainable spatial reasoning.\n\n## Features\n- **3D Scene Management**: Create and manage complex 3D interior layouts\n- **Asset Processing Pipeline**: Automatic optimization with KTX2, Draco, and Meshopt variants\n- **Real-time Collaboration**: Server-Sent Events for live scene updates\n- **Optimistic Locking**: If-Match headers for safe concurrent editing\n- **Category-based Organization**: Enhanced scene manifests with filtering capabilities\n- **Processing Queue**: Background asset processing with status tracking\n\n## Authentication\nAll endpoints require Bearer token authentication unless explicitly marked as public.\n\n## Rate Limiting\n- General API: 100 requests per 15 minutes per IP\n- Authentication: 5 attempts per 15 minutes per IP\n\n## Versioning\nScene operations support optimistic locking via If-Match headers for safe concurrent editing.",
    "version": "1.0.0",
    "contact": {
      "name": "Lumea Development Team",
      "url": "https://lumea.dev",
      "email": "support@lumea.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://192.168.1.10:3000",
      "description": "Network Access Server"
    },
    {
      "url": "http://192.168.1.10:3000",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}